+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.5.0
 * @akashic/akashic-engine@3.3.0 に追従
+* @akashic/game-configuration@1.2.0 に追従
 
 ## 2.4.3
 * スナップショットで外部からリセットした場合に、`TickBuffer#knownLatestAge` が更新されない問題を修正